arrayTest fix

+++ tests/arrayTest.c
@@ -2,9 +2,9 @@ int main(){
     int s = 10;
     int arr[10];
     int k;
-    float brr[10];
+    int brr[10];
     char crr[50];
-    arr[2] = 19.0;
+    arr[2] = 19;
     k = arr[s-2];
     crr[3] = 'q';
 }
